Move `VectorSink`, `StreamSink` into core
This commit moves `VectorSink` and `StreamSink` into core, leaving behind a fair
bit extra of the sink traits. Sink is our most complicated notion in vector by
far and fairly self-referential in its implementation. Getting it teased into
core will take some doing and will, very likely, need to be driven by need.

Future work in this area will be done one piece at a time. There's simply too
much code to move in bulk.

REF #7150
