+++ src/sinks/mod.rs
@@ -1,7 +1,5 @@
-use crate::event::Event;
-use futures::{future::BoxFuture, Sink, Stream, StreamExt};
+use futures::future::BoxFuture;
 use snafu::Snafu;
-use std::fmt;
 
 pub mod util;
 
@@ -68,10 +66,7 @@ pub mod statsd;
 #[cfg(feature = "sinks-vector")]
 pub mod vector;
 
-pub enum VectorSink {
-    Sink(Box<dyn Sink<Event, Error = ()> + Send + Unpin>),
-    Stream(Box<dyn util::StreamSink + Send>),
-}
+pub use vector_core::sink::VectorSink;
 
 pub type Healthcheck = BoxFuture<'static, crate::Result<()>>;
 
@@ -93,29 +88,4 @@ pub enum BuildError {
 pub enum HealthcheckError {
     #[snafu(display("Unexpected status: {}", status))]
     UnexpectedStatus { status: ::http::StatusCode },
-}
-
-impl VectorSink {
-    pub async fn run<S>(mut self, input: S) -> Result<(), ()>
-    where
-        S: Stream<Item = Event> + Send,
-    {
-        match self {
-            Self::Sink(sink) => input.map(Ok).forward(sink).await,
-            Self::Stream(ref mut s) => s.run(Box::pin(input)).await,
-        }
-    }
-
-    pub fn into_sink(self) -> Box<dyn Sink<Event, Error = ()> + Send + Unpin> {
-        match self {
-            Self::Sink(sink) => sink,
-            _ => panic!("Failed type coercion, {:?} is not a Sink", self),
-        }
-    }
-}
-
-impl fmt::Debug for VectorSink {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("VectorSink").finish()
-    }
 }
